29052024_Suba

+++ src/app/pages/category/category.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { HttpService } from 'src/app/http.service';
 import { firstValueFrom } from 'rxjs';
 import {ActivatedRoute, Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-category',
@@ -39,20 +40,33 @@ export class CategoryComponent implements OnInit {
     console.log(data);
   }
 
+  // async deleteCategory(id: number): Promise<void> {
+  //   try {
+      
+  //     await this.HttpService.deleteCategory(id).toPromise();
+  //     console.log(`Category with ID ${id} deleted successfully.`);
+  //     this.fetchCategories();
+      
+  //   } catch (error) {
+  //     console.error('Error deleting category:', error);
+      
+  //   }
+  // }
+
   async deleteCategory(id: number): Promise<void> {
-    try {
-      
-      await this.HttpService.deleteCategory(id).toPromise();
-      console.log(`Category with ID ${id} deleted successfully.`);
-      this.fetchCategories();
-      
-    } catch (error) {
-      console.error('Error deleting category:', error);
-      
+    const isConfirmed = await this.showConfirmation();
+  
+    if (isConfirmed) {
+      try {
+        // Your delete logic here
+        await this.HttpService.deleteCategory(id).toPromise();
+        console.log(`Category with ID ${id} deleted successfully.`);
+        this.fetchCategories();
+      } catch (error) {
+        console.error('Error deleting category:', error);
+      }
     }
   }
-
-
 
   
   // async fetchCategories(): Promise<void> {
@@ -109,5 +123,18 @@ export class CategoryComponent implements OnInit {
 //     }
 //   );
 // }
+async showConfirmation(): Promise<boolean> {
+  const result = await Swal.fire({
+    title: 'Are you sure?',
+    text: 'You will not be able to recover this record!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  });
+
+  return result.isConfirmed;
+}
 
 }
